# demo/demo7/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Advisement Appointment API",
    "version": "1.0.0",
    "description": "A small local API for scheduling BYU advisement appointments. This document is the machine-readable contract for the current demo version. There is no deprecation notice for version 1.0.0. Clients should treat a changed major version as a potentially breaking contract change."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local demo server"
    }
  ],
  "tags": [
    {
      "name": "Appointments",
      "description": "Advisement appointment operations"
    }
  ],
  "paths": {
    "/advisement/appointments": {
      "post": {
        "tags": ["Appointments"],
        "operationId": "createAdvisementAppointment",
        "summary": "Schedule an advisement appointment",
        "description": "Creates one appointment for the authenticated student. The caller must be authorized to schedule for the student_id in the request. If Idempotency-Key is supplied, retrying the same request with the same key replays the original response without creating another appointment. Reusing that key with different request data is rejected. When rate limiting is enabled, clients must wait for Retry-After before retrying a 429 response. Appointment slots are exclusive per advisor and starts_at.",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentRequest"
              },
              "examples": {
                "academicAdvising": {
                  "summary": "Schedule academic advising",
                  "value": {
                    "student_id": "123456789",
                    "advisor_id": "111111111",
                    "appointment_type": "academic_advising",
                    "starts_at": "2026-10-08T17:00:00-07:00"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment created or the original response replayed for an idempotent retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                },
                "examples": {
                  "created": {
                    "value": {
                      "id": 1,
                      "student_id": "123456789",
                      "advisor_id": "111111111",
                      "appointment_type": "academic_advising",
                      "starts_at": "2026-10-08T17:00:00-07:00"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request body is invalid JSON, contains unknown fields, has multiple JSON values, is missing a required field, or uses an invalid timestamp or appointment type.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "invalidJson": {
                    "value": {
                      "type": "http://localhost:8080/problems/invalid-json",
                      "title": "Invalid JSON request body",
                      "status": 400,
                      "code": "INVALID_JSON",
                      "detail": "The request body must contain valid JSON with only supported fields."
                    }
                  },
                  "invalidAppointment": {
                    "value": {
                      "type": "http://localhost:8080/problems/invalid-appointment",
                      "title": "Invalid appointment",
                      "status": 400,
                      "code": "INVALID_APPOINTMENT",
                      "detail": "starts_at must be a valid RFC3339 timestamp"
                    }
                  },
                  "multipleJsonValues": {
                    "value": {
                      "type": "http://localhost:8080/problems/multiple-json-values",
                      "title": "Multiple JSON values",
                      "status": 400,
                      "code": "MULTIPLE_JSON_VALUES",
                      "detail": "The request body must contain exactly one JSON object."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AppointmentNotPermitted"
          },
          "409": {
            "description": "The requested slot is unavailable, or an idempotency key was reused with different request data.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "slotUnavailable": {
                    "value": {
                      "type": "http://localhost:8080/problems/appointment-slot-unavailable",
                      "title": "Appointment slot unavailable",
                      "status": 409,
                      "code": "APPOINTMENT_SLOT_UNAVAILABLE",
                      "detail": "The requested appointment slot is no longer available."
                    }
                  },
                  "idempotencyKeyReused": {
                    "value": {
                      "type": "http://localhost:8080/problems/idempotency-key-reused",
                      "title": "Idempotency key already used",
                      "status": 409,
                      "code": "IDEMPOTENCY_KEY_REUSED",
                      "detail": "The idempotency key was already used with different request data."
                    }
                  }
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimitExceeded"
          },
          "500": {
            "description": "The server could not complete the operation. The client should not assume that the appointment was not created; use the idempotency key when retrying.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "http://localhost:8080/problems/internal-server-error",
                  "title": "Internal server error",
                  "status": 500,
                  "code": "INTERNAL_SERVER_ERROR",
                  "detail": "The server could not complete the request."
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "description": "A client-generated key for one intended create operation. Reuse the same key and identical request data when retrying after an uncertain response.",
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "example": "demo-retry-1"
      }
    },
    "responses": {
      "AuthenticationRequired": {
        "description": "The request did not include a valid bearer token.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "http://localhost:8080/problems/authentication-required",
              "title": "Authentication required",
              "status": 401,
              "code": "AUTHENTICATION_REQUIRED",
              "detail": "Provide a valid bearer token."
            }
          }
        }
      },
      "AppointmentNotPermitted": {
        "description": "The caller is authenticated but is not allowed to schedule this appointment.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "http://localhost:8080/problems/appointment-not-permitted",
              "title": "Appointment not permitted",
              "status": 403,
              "code": "APPOINTMENT_NOT_PERMITTED",
              "detail": "The caller is not allowed to schedule this appointment."
            }
          }
        }
      },
      "RateLimitExceeded": {
        "description": "The client exceeded the configured rate limit. This behavior is enabled for the demo with DEMO_RATE_LIMIT.",
        "headers": {
          "Retry-After": {
            "description": "Minimum number of seconds to wait before retrying.",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 10
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "http://localhost:8080/problems/rate-limit-exceeded",
              "title": "Rate limit exceeded",
              "status": 429,
              "code": "RATE_LIMIT_EXCEEDED",
              "detail": "Too many requests from this client. Retry after 10 seconds."
            }
          }
        }
      }
    },
    "schemas": {
      "CreateAppointmentRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "student_id",
          "advisor_id",
          "appointment_type",
          "starts_at"
        ],
        "properties": {
          "student_id": {
            "type": "string",
            "description": "BYU ID of the student receiving advising.",
            "example": "123456789"
          },
          "advisor_id": {
            "type": "string",
            "description": "BYU ID of the advisor.",
            "example": "111111111"
          },
          "appointment_type": {
            "type": "string",
            "description": "The kind of advising appointment.",
            "enum": ["academic_advising"],
            "example": "academic_advising"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "description": "Appointment start time as an RFC3339 timestamp including its UTC offset or Z suffix.",
            "example": "2026-10-08T17:00:00-07:00"
          }
        }
      },
      "Appointment": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "student_id",
          "advisor_id",
          "appointment_type",
          "starts_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "student_id": {
            "type": "string",
            "example": "123456789"
          },
          "advisor_id": {
            "type": "string",
            "example": "111111111"
          },
          "appointment_type": {
            "type": "string",
            "enum": ["academic_advising"],
            "example": "academic_advising"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-08T17:00:00-07:00"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "required": ["type", "title", "status", "code", "detail"],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "description": "A URI identifying the problem type and its client guidance."
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "demo-token",
        "description": "Use a demo bearer token. The demo-student identity may create an appointment only for student_id 123456789."
      }
    }
  }
}